requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0  # For LoRA/QLoRA
bitsandbytes>=0.41.0  # For quantization

# Training & Optimization
trl>=0.7.0  # Transformer Reinforcement Learning
scipy>=1.11.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Evaluation & Metrics
evaluate>=0.4.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
wandb>=0.16.0  # Experiment tracking
tensorboard>=2.15.0
tqdm>=4.66.0
pandas>=2.1.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: For specific model types
einops>=0.7.0
flash-attn>=2.3.0  # Optional: for faster attention (requires specific GPU)

